Create 07_personalization_of_online_store_offers
The online store "In One Click" sells various goods: for children, for the home, small household appliances, cosmetics and even food. The store's report for the previous period showed that customer activity began to decline. Attracting new customers is no longer so effective: most of the target audience already knows about the store. A possible solution is to retain the activity of regular customers. This can be done with the help of personalized offers.